Fix problem with tabs
